added buttons and output, styling added
Buttons and output were added the challenge currently is that i want the calculator to be a
seperate div standing alone. Originally in the old design there was not background that had to
stand out from the body. I am going to separate the output from the buttons into to seperate divs
and style them accordingly.