Logic added for new digits
Logic was added to the ADD_DIGIT action in the reducer that prevents mulitple zeros from being input if the
only operand is a zero. Logic was added to ensure that if the dot button is pressed and the only value is a
zero that operand will retain the zero before the decimal.